// src/store/types/profile.ts
export interface ProfileState {
  fullName: string;
  email: string;
  profession: string;
  bio: string;
  phone: string;
  birthDate: string;
  postalCode: string;
  street: string;
  neighborhood: string;
  number: string;
  city: string;
  state: string;
  currentPassword: string;
  newPassword: string;
  confirmPassword: string;
  isFetchingAddress: boolean;
}